+++ lib/trailblazer/operation/trace.rb
@@ -1,4 +1,4 @@
-require 'delegate'
+require "delegate"
 require "trailblazer/developer"
 
 module Trailblazer
@@ -13,7 +13,7 @@ module Trailblazer
 
         result = Railway::Result(signal, ctx) # redundant with PublicCall::call.
 
-        Result.new(result, stack.to_a)
+        Result.new(result, stack)
       end
 
       # `Operation::trace` is included for simple tracing of the flow.
@@ -22,12 +22,24 @@ module Trailblazer
       # @public
       #
       #   Operation.trace(params, current_user: current_user).wtf
+      # TODO: remove in 0.11.0.
       def trace(options)
+        Activity::Deprecate.warn Trace.find_caller_location_for_deprecated, %(Using `Operation.trace` is deprecated and will be removed in {trailblazer-operation-0.11.0}.
+  Please use `#{self}.wtf?` as documented here: https://trailblazer.to/2.1/docs/trailblazer#trailblazer-developer-wtf-)
+
         Trace.(self, options)
       end
 
       def wtf?(options)
         call_with_public_interface(options, {}, invoke_class: Developer::Wtf)
+      end
+
+      # TODO: remove in 0.11.0.
+      def self.find_caller_location_for_deprecated
+        our_caller_locations = caller_locations.to_a
+        caller_location = our_caller_locations.reverse.find { |line| line.to_s =~ /operation\/trace/ }
+
+        _caller_location = our_caller_locations[our_caller_locations.index(caller_location)+1]
       end
 
       # Presentation of the traced stack via the returned result object.
@@ -38,12 +50,14 @@ module Trailblazer
           @stack = stack
         end
 
-        # TODO: deprecate!
+        # TODO: remove in 0.11.0.
         def wtf
+          Activity::Deprecate.warn Trace.find_caller_location_for_deprecated, %(Using `result.wtf?` is deprecated. Please use `#{@stack.to_a[0].task}.wtf?` and have a nice day.)
+
           Developer::Trace::Present.(@stack)
         end
 
-        # TODO: deprecate!
+        # TODO: remove in 0.11.0.
         def wtf?
           puts wtf
         end