+++ test/trace_test.rb
@@ -12,7 +12,31 @@ class TraceTest < Minitest::Spec
     step ->(options, **) { options[:c] = true }, id: "Create.task.c"
     step ->(_options, params:, **) { params.any? }, id: "Create.task.params"
   end
-  # raise Create["__task_wraps__"].inspect
+
+  it "deprecates {result.wtf?} and {Operation.trace}" do
+    output, warning = capture_io do
+      result = Create.trace({a_return: true, params: {}})
+      result.wtf?
+    end
+    line_no = __LINE__ - 2
+
+    assert_equal warning, %([Trailblazer] #{File.realpath(__FILE__)}:#{line_no - 1} Using `Operation.trace` is deprecated and will be removed in {trailblazer-operation-0.11.0}.
+  Please use `TraceTest::Create.wtf?` as documented here: https://trailblazer.to/2.1/docs/trailblazer#trailblazer-developer-wtf-
+[Trailblazer] #{File.realpath(__FILE__)}:#{line_no} Using `result.wtf?` is deprecated. Please use `TraceTest::Create.wtf?` and have a nice day.
+)
+    assert_equal output, %(TraceTest::Create
+|-- Start.default
+|-- Create.task.a
+|-- MyNested
+|   |-- Start.default
+|   |-- B.task.b
+|   |-- B.task.e
+|   `-- End.success
+|-- Create.task.c
+|-- Create.task.params
+`-- End.failure
+)
+  end
 
   it "allows using low-level Operation::Trace" do
     result = Trailblazer::Operation::Trace.(